+++ material-app/build-logic/convention/src/main/kotlin/AndroidApplicationConventionPlugin.kt
@@ -11,6 +11,8 @@ class AndroidApplicationConventionPlugin : Plugin<Project> {
             extensions.configure(CommonExtension::class.java) {
                 configureAndroidCommon(this)
             }
+
+            registerBuildConventions()
         }
     }
 }