+++ material-app/samples/actions/src/main/java/dev/qtremors/material/samples/actions/ActionSamples.kt
@@ -58,8 +58,8 @@ import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.OutlinedIconButton
 import androidx.compose.material3.OutlinedIconToggleButton
 import androidx.compose.material3.SmallFloatingActionButton
+import androidx.compose.material3.SplitButton
 import androidx.compose.material3.SplitButtonDefaults
-import androidx.compose.material3.SplitButtonLayout
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
@@ -438,7 +438,7 @@ fun RejectedActionButton(modifier: Modifier = Modifier) {
 fun SplitButtonsSample(modifier: Modifier = Modifier) {
     var expanded by remember { mutableStateOf(false) }
     Column(modifier) {
-        SplitButtonLayout(
+        SplitButton(
             leadingButton = {
                 SplitButtonDefaults.LeadingButton(onClick = {}) {
                     Icon(Icons.Default.Share, contentDescription = null)